[C2-6008] Payment tools.
