+++ src/client/index.ts
@@ -1,4 +1,2 @@
 import { YunoClient } from './YunoClient';
-
 export { YunoClient };
-export * from './types';
