+++ src/customers/index.ts
@@ -1,0 +1,12 @@
+import { Tool } from "../shared/types";
+import { customerCreateTool } from "./toolCreate";
+import { customerRetrieveTool } from "./toolRetrieve";
+import { customerRetrieveByExternalIdTool } from "./toolRetrieveByExternalId";
+import { customerUpdateTool } from "./toolUpdate";
+
+export const customerTools: Tool[] = [
+  customerCreateTool,
+  customerRetrieveTool,
+  customerRetrieveByExternalIdTool,
+  customerUpdateTool,
+];
