+++ src/shared/types/index.ts
@@ -1,0 +1,4 @@
+export * from './customer';
+export * from './common';
+export * from './checkout';
+export * from './payment';
